obligation revamp
