+++ Advanced/obligation-cordapp/CordaTraining.postman_collection.json
@@ -1,7 +1,7 @@
 {
 	"info": {
 		"_postman_id": "a15974fd-2343-87ea-e9a1-b6b228aa8093",
-		"name": "Training Corda App",
+		"name": "Obligation Corda App",
 		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
 	},
 	"item": [
