+++ Advanced/obligation-cordapp/build.gradle
@@ -91,8 +91,8 @@ dependencies {
     cordaRuntime "$corda_release_group:corda:$corda_release_version"
 
     // CorDapp dependencies.
-    cordapp project(":workflows-kotlin")
-    cordapp project(":contracts-kotlin")
+    cordapp project(":workflows")
+    cordapp project(":contracts")
     cordapp "$corda_release_group:corda-finance-contracts:$corda_release_version"
     cordapp "$corda_release_group:corda-finance-workflows:$corda_release_version"
     cordapp "$corda_release_group:corda-confidential-identities:$corda_release_version"
@@ -119,13 +119,13 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
         projectCordapp {
             deploy = false
         }
-        cordapp project(':contracts-kotlin')
-        cordapp project(':workflows-kotlin')
+        cordapp project(':contracts')
+        cordapp project(':workflows')
         cordapp("$corda_release_group:corda-finance-contracts:$corda_release_version")
         cordapp("$corda_release_group:corda-finance-workflows:$corda_release_version")
         cordapp("$corda_release_group:corda-confidential-identities:$corda_release_version")
         rpcUsers = [[ user: "user1", "password": "password", "permissions": ["ALL"]]]
-	runSchemaMigration = true
+        runSchemaMigration = true
     }
     node {
         name "O=Notary,L=London,C=GB"
