+++ Advanced/obligation-cordapp/clients/build.gradle
@@ -46,8 +46,8 @@ dependencies {
     compile "org.springframework.boot:spring-boot-starter-log4j2:$spring_boot_version"
     compile "org.apache.logging.log4j:log4j-web:$log4j_version"
 
-    cordapp project(":workflows-kotlin")
-    cordapp project(":contracts-kotlin")
+    cordapp project(":workflows")
+    cordapp project(":contracts")
 
 }
 
@@ -63,19 +63,19 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 
 task runPartyAServer(type: JavaExec, dependsOn: jar) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'net.corda.training.server.ServerKt'
+    main = 'net.corda.samples.obligation.server.ServerKt'
     args '--server.port=10009', '--config.rpc.host=localhost', '--config.rpc.port=10008', '--config.rpc.username=user1', '--config.rpc.password=password'
 }
 
 task runPartyBServer(type: JavaExec, dependsOn: jar) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'net.corda.training.server.ServerKt'
+    main = 'net.corda.samples.obligation.server.ServerKt'
     args '--server.port=10012', '--config.rpc.host=localhost', '--config.rpc.port=10011', '--config.rpc.username=user1', '--config.rpc.password=password'
 }
 
 task runPartyCServer(type: JavaExec, dependsOn: jar) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'net.corda.training.server.ServerKt'
+    main = 'net.corda.samples.obligation.server.ServerKt'
     args '--server.port=10015', '--config.rpc.host=localhost', '--config.rpc.port=10014', '--config.rpc.username=user1', '--config.rpc.password=password'
 }
