--- Advanced/obligation-cordapp/clients/src/main/kotlin/net/corda/training/server/Server.kt
@@ -1,37 +1,0 @@
-package net.corda.training.server
-
-import net.corda.client.jackson.JacksonSupport
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.Banner
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.context.annotation.Bean
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
-
-/**
- * A Spring Boot application.
- */
-
-@SpringBootApplication
-open class Server {
-    /**
-     * Spring Bean that binds a Corda Jackson object-mapper to HTTP message types used in Spring.
-     */
-    @Bean
-    open fun mappingJackson2HttpMessageConverter(@Autowired rpcConnection: NodeRPCConnection): MappingJackson2HttpMessageConverter {
-        val mapper = JacksonSupport.createDefaultMapper(rpcConnection.proxy)
-        val converter = MappingJackson2HttpMessageConverter()
-        converter.objectMapper = mapper
-        return converter
-    }
-}
-
-/**
- * Starts our Spring Boot application.
- */
-fun main(args: Array<String>) {
-    val app = SpringApplication(Server::class.java)
-    app.setBannerMode(Banner.Mode.OFF)
-    app.isWebEnvironment = true
-    app.run(*args)
-}
