--- Advanced/obligation-cordapp/contracts-kotlin/src/main/kotlin/net/corda/training/states/IOUState.kt
@@ -1,33 +1,0 @@
-package net.corda.training.states
-
-import net.corda.core.contracts.*
-import net.corda.training.contracts.IOUContract
-import net.corda.core.identity.Party
-import java.util.*
-
-/**
- * This is where you'll add the definition of your state object. Look at the unit tests in [IOUStateTests] for
- * instructions on how to complete the [IOUState] class.
- *
- * Remove the "val data: String = "data" property before starting the [IOUState] tasks.
- */
-@BelongsToContract(IOUContract::class)
-data class IOUState(val amount: Amount<Currency>,
-                    val lender: Party,
-                    val borrower: Party,
-                    val paid: Amount<Currency> = Amount(0, amount.token),
-                    override val linearId: UniqueIdentifier = UniqueIdentifier()): LinearState {
-    /**
-     *  This property holds a list of the nodes which can "use" this state in a valid transaction. In this case, the
-     *  lender or the borrower.
-     */
-    override val participants: List<Party> get() = listOf(lender, borrower)
-
-    /**
-     * Helper methods for when building transactions for settling and transferring IOUs.
-     * - [pay] adds an amount to the paid property. It does no validation.
-     * - [withNewLender] creates a copy of the current state with a newly specified lender. For use when transferring.
-     */
-    fun pay(amountToPay: Amount<Currency>) = copy(paid = paid.plus(amountToPay))
-    fun withNewLender(newLender: Party) = copy(lender = newLender)
-}
