--- Advanced/obligation-cordapp/contracts-kotlin/src/test/kotlin/net/corda/training/contracts/ContractTests.kt
@@ -1,13 +1,0 @@
-package net.corda.training.contracts
-
-import net.corda.testing.node.MockServices
-import org.junit.Test
-
-class ContractTests {
-    private val ledgerServices = MockServices()
-
-    @Test
-    fun `dummy test`() {
-
-    }
-}
