+++ Advanced/obligation-cordapp/gradle.properties
@@ -1,4 +1,4 @@
 name=Test
-group=com.template
+group=com.obligation
 version=0.1
 kotlin.incremental=false
