+++ Advanced/obligation-cordapp/settings.gradle
@@ -1,3 +1,3 @@
-include 'workflows-kotlin'
-include 'contracts-kotlin'
+include 'workflows'
+include 'contracts'
 include 'clients'
