--- Advanced/obligation-cordapp/workflows-kotlin/build.gradle
@@ -1,69 +1,0 @@
-apply plugin: 'net.corda.plugins.cordapp'
-apply plugin: 'net.corda.plugins.quasar-utils'
-
-cordapp {
-    targetPlatformVersion corda_platform_version
-    minimumPlatformVersion corda_platform_version
-    workflow {
-        name "Template Flows"
-        vendor "Corda Open Source"
-        licence "Apache License, Version 2.0"
-        versionId 1
-    }
-}
-
-sourceSets {
-    main {
-        resources {
-            srcDir rootProject.file("config/dev")
-        }
-    }
-    test {
-        resources {
-            srcDir rootProject.file("config/test")
-        }
-    }
-    integrationTest {
-        kotlin {
-            compileClasspath += main.output + test.output
-            runtimeClasspath += main.output + test.output
-            srcDir file('src/integrationTest/kotlin')
-        }
-    }
-}
-
-configurations {
-    integrationTestCompile.extendsFrom testCompile
-    integrationTestRuntime.extendsFrom testRuntime
-}
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-    testCompile "junit:junit:$junit_version"
-
-    // Corda dependencies.
-    cordaCompile "$corda_core_release_group:corda-core:$corda_core_release_version"
-    cordaCompile "$corda_core_release_group:corda-finance-contracts:$corda_release_version"
-    cordaCompile "$corda_core_release_group:corda-finance-workflows:$corda_release_version"
-
-    cordaRuntime "$corda_release_group:corda:$corda_release_version"
-    
-    //finance 
-    cordaCompile "$corda_release_group:corda-finance-contracts:$corda_release_version"
-    cordaCompile "$corda_release_group:corda-finance-workflows:$corda_release_version"
-
-    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-
-    // CorDapp dependencies.
-    cordapp project(":contracts-kotlin")
-    cordapp "$corda_release_group:corda-finance-contracts:$corda_release_version"
-    cordapp "$corda_release_group:corda-finance-workflows:$corda_release_version"
-    cordapp("$corda_release_group:corda-confidential-identities:$corda_release_version")
-
-}
-
-task integrationTest(type: Test, dependsOn: []) {
-    testClassesDirs = sourceSets.integrationTest.output.classesDirs
-    classpath = sourceSets.integrationTest.runtimeClasspath
-}
