+++ Advanced/obligation-cordapp/workflows/src/main/kotlin/net/corda/samples/obligation/flows/IOUIssueFLow.kt
@@ -1,0 +1,79 @@
+package net.corda.samples.obligation.flows
+
+
+import co.paralleluniverse.fibers.Suspendable
+import net.corda.core.contracts.Command
+import net.corda.samples.obligation.states.IOUState
+import net.corda.core.contracts.requireThat
+import net.corda.core.flows.*
+import net.corda.core.transactions.SignedTransaction
+import net.corda.core.transactions.TransactionBuilder
+import net.corda.samples.obligation.contract.IOUContract
+
+/**
+ * This is the flow which handles issuance of new IOUs on the ledger.
+ * Gathering the counterparty's signature is handled by the [CollectSignaturesFlow].
+ * Notarisation (if required) and commitment to the ledger is handled by the [FinalityFlow].
+ * The flow returns the [SignedTransaction] that was committed to the ledger.
+ */
+@InitiatingFlow
+@StartableByRPC
+class IOUIssueFlow(val state: IOUState): FlowLogic<SignedTransaction>() {
+    @Suspendable
+    override fun call(): SignedTransaction {
+
+        // Step 1. Get a reference to the notary service on our network and our key pair.
+        // Note: ongoing work to support multiple notary identities is still in progress.
+
+        /**
+         *  METHOD 1: Take first notary on network, WARNING: use for test, non-prod environments, and single-notary networks only!*
+         *  METHOD 2: Explicit selection of notary by CordaX500Name - argument can by coded in flow or parsed from config (Preferred)
+         *
+         *  * - For production you always want to use Method 2 as it guarantees the expected notary is returned.
+         */
+        val notary = serviceHub.networkMapCache.notaryIdentities.single() // METHOD 1
+        // val notary = serviceHub.networkMapCache.getNotary(CordaX500Name.parse("O=Notary,L=London,C=GB")) // METHOD 2
+
+        // Step 2. Create a new issue command.
+        // Remember that a command is a CommandData object and a list of CompositeKeys
+        val issueCommand = Command(IOUContract.Commands.Issue(), state.participants.map { it.owningKey })
+
+        // Step 3. Create a new TransactionBuilder object.
+        val builder = TransactionBuilder(notary = notary)
+
+        // Step 4. Add the iou as an output state, as well as a command to the transaction builder.
+        builder.addOutputState(state, IOUContract.IOU_CONTRACT_ID)
+        builder.addCommand(issueCommand)
+
+        // Step 5. Verify and sign it with our KeyPair.
+        builder.verify(serviceHub)
+        val ptx = serviceHub.signInitialTransaction(builder)
+
+        val sessions = (state.participants - ourIdentity).map { initiateFlow(it) }.toSet()
+        // Step 6. Collect the other party's signature using the SignTransactionFlow.
+        val stx = subFlow(CollectSignaturesFlow(ptx, sessions))
+
+        // Step 7. Assuming no exceptions, we can now finalise the transaction.
+        return subFlow(FinalityFlow(stx, sessions))
+    }
+}
+
+/**
+ * This is the flow which signs IOU issuances.
+ * The signing is handled by the [SignTransactionFlow].
+ */
+@InitiatedBy(IOUIssueFlow::class)
+class IOUIssueFlowResponder(val flowSession: FlowSession): FlowLogic<SignedTransaction>() {
+
+    @Suspendable
+    override fun call(): SignedTransaction {
+        val signedTransactionFlow = object : SignTransactionFlow(flowSession) {
+            override fun checkTransaction(stx: SignedTransaction) = requireThat {
+                val output = stx.tx.outputs.single().data
+                "This must be an IOU transaction" using (output is IOUState)
+            }
+        }
+        val txWeJustSignedId = subFlow(signedTransactionFlow)
+        return subFlow(ReceiveFinalityFlow(otherSideSession = flowSession, expectedTxId = txWeJustSignedId.id))
+    }
+}
